добавит заглушку при отсутсвии предложений
