+++ project/src/pages/main-screen/main-empty-screen.tsx
@@ -1,15 +1,23 @@
-function MainEmptyScreen(): JSX.Element {
+type MainEmptyScreenProps = {
+  currentCity: string
+}
+
+function MainEmptyScreen({currentCity}: MainEmptyScreenProps): JSX.Element {
   return (
-    <div className="cities__places-container cities__places-container--empty container">
-      <section className="cities__no-places">
-        <div className="cities__status-wrapper tabs__content">
-          <b className="cities__status">No places to stay available</b>
-          <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
-        </div>
-      </section>
-      <div className="cities__right-section"></div>
+    <div className="cities">
+      <div className="cities__places-container cities__places-container--empty container">
+        <section className="cities__no-places">
+          <div className="cities__status-wrapper tabs__content">
+            <b className="cities__status">No places to stay available</b>
+            <p className="cities__status-description">We could not find any property available at the moment in {currentCity}</p>
+          </div>
+        </section>
+        <div className="cities__right-section"></div>
+      </div>
     </div>
   );
 }
 
 export default MainEmptyScreen;
+
+
